Separate the agent logic
